framer updates for animation on the landing page.
